+++ integrationlink/update.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/Scalingo/cli/config"
 	"github.com/Scalingo/cli/io"
+	"github.com/Scalingo/cli/utils"
 	"github.com/Scalingo/go-scalingo"
 )
 
@@ -20,7 +21,13 @@ func Update(app string, params scalingo.SCMRepoLinkUpdateParams) error {
 
 	_, err = c.SCMRepoLinkUpdate(app, params)
 	if err != nil {
-		return errgo.Notef(err, "fail to update integration link")
+		if !utils.IsPaymentRequiredAndFreeTrialExceededError(err) {
+			return errgo.Notef(err, "fail to update integration link")
+		}
+
+		return utils.AskAndStopFreeTrial(c, func() error {
+			return Update(app, params)
+		})
 	}
 
 	io.Statusf("Your app '%s' integration link has been updated.\n", app)